requirements.txt
Django>=4.1,<4.2
psycopg2-binary>=2.9,<3.0
django-environ>=0.7,<1.0
djangorestframework>=3.15.0
djangorestframework-simplejwt
requests